+++ react/__mocks__/vtex.totalizer-translator.tsx
@@ -1,0 +1,7 @@
+import React, { FunctionComponent } from 'react'
+
+export const TranslateTotalizer: FunctionComponent<{ id: string }> = ({
+  id,
+}) => {
+  return <div>{id}</div>
+}