+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Using `SkeletonPiece` component from `vtex.my-account-commons`.
+- Using `TranslateTotalizer` component from `vtex.totalizer-translator`.
 
 ### Fixed